Refactor command handling into CommandProcessor class Mark async tests as async individually Fix async test fixtures
