+++ .pre-commit-config.yaml
@@ -9,7 +9,7 @@ repos:
   - id: debug-statements
   - id: check-ast
 - repo: https://github.com/psf/black
-  rev: 20.8b1
+  rev: 22.3.0
   hooks:
   - id: black
 - repo: https://gitlab.com/pycqa/flake8
