+++ pyotgw/protocol.py
@@ -3,9 +3,9 @@
 import asyncio
 import logging
 import re
-from asyncio.queues import QueueFull
 
 from pyotgw import vars as v
+from pyotgw.commandprocessor import CommandProcessor
 from pyotgw.messageprocessor import MessageProcessor
 
 _LOGGER = logging.getLogger(__name__)
@@ -26,15 +26,17 @@ class OpenThermProtocol(asyncio.Protocol):
         """Initialise the protocol object."""
         self.transport = None
         self.loop = loop
-        self._active = False
-        self._cmd_lock = asyncio.Lock()
-        self._cmdq = asyncio.Queue()
         self._readbuf = b""
         self._received_lines = 0
         self.activity_callback = activity_callback
+        self.command_processor = CommandProcessor(
+            self,
+            status_manager,
+            loop,
+        )
         self._connected = False
         self.message_processor = MessageProcessor(
-            self,
+            self.command_processor,
             status_manager,
             loop,
         )
@@ -53,13 +55,11 @@ class OpenThermProtocol(asyncio.Protocol):
         """
         if self.active:
             _LOGGER.error("Disconnected: %s", exc)
-        self._active = False
+        self._received_lines = 0
         self._connected = False
-        self.transport.close()
+        self.command_processor.clear_queue()
         self.message_processor.connection_lost()
         self.status_manager.reset()
-        while not self._cmdq.empty():
-            self._cmdq.get_nowait()
 
     @property
     def connected(self):
@@ -84,7 +84,6 @@ class OpenThermProtocol(asyncio.Protocol):
         Perform line buffering and call line_received() with complete
         lines.
         """
-        self._active = True
         # DIY line buffering...
         newline = b"\r\n"
         eot = b"\x04"
@@ -121,108 +120,19 @@ class OpenThermProtocol(asyncio.Protocol):
             self._received_lines = 0
             _LOGGER.debug("Ignoring line: %s", line)
         else:
-            try:
-                self._cmdq.put_nowait(line)
-                _LOGGER.debug(
-                    "Added line %d to command queue. Queue size: %d",
-                    self._received_lines,
-                    self._cmdq.qsize(),
-                )
-            except QueueFull:
-                _LOGGER.error("Queue full, discarded message: %s", line)
-
-    async def issue_cmd(self, cmd, value, retry=3):
-        """
-        Issue a command, then await and return the return value.
-
-        This method is a coroutine
-        """
-        async with self._cmd_lock:
-            if not self.connected:
-                _LOGGER.debug("Serial transport closed, not sending command %s", cmd)
-                return
-            while not self._cmdq.empty():
-                _LOGGER.debug(
-                    "Clearing leftover message from command queue: %s",
-                    await self._cmdq.get(),
-                )
-            if isinstance(value, float):
-                value = f"{value:.2f}"
-            _LOGGER.debug("Sending command: %s with value %s", cmd, value)
-            self.transport.write(f"{cmd}={value}\r\n".encode("ascii"))
-            if cmd == v.OTGW_CMD_REPORT:
-                expect = fr"^{cmd}:\s*([A-Z]{{2}}|{value}=[^$]+)$"
-            # OTGW_CMD_CONTROL_HEATING_2 and OTGW_CMD_CONTROL_SETPOINT_2 do not adhere
-            # to the standard response format (<cmd>: <value>) at the moment, but report
-            # only the value. This will likely be fixed in the future, so we support
-            # both formats.
-            elif cmd in (
-                v.OTGW_CMD_CONTROL_HEATING_2,
-                v.OTGW_CMD_CONTROL_SETPOINT_2,
-            ):
-                expect = fr"^(?:{cmd}:\s*)?(0|1|[0-9]+\.[0-9]{{2}}|[A-Z]{{2}})$"
-            else:
-                expect = fr"^{cmd}:\s*([^$]+)$"
-
-            async def send_again(err):
-                """Resend the command."""
-                nonlocal retry
-                _LOGGER.warning("Command %s failed with %s, retrying...", cmd, err)
-                retry -= 1
-                self.transport.write(f"{cmd}={value}\r\n".encode("ascii"))
-
-            async def process(msg):
-                """Process a possible response."""
-                _LOGGER.debug("Got possible response for command %s: %s", cmd, msg)
-                if msg in v.OTGW_ERRS:
-                    # Some errors appear by themselves on one line.
-                    if retry == 0:
-                        raise v.OTGW_ERRS[msg]
-                    await send_again(msg)
-                    return
-                if cmd == v.OTGW_CMD_MODE and value == "R":
-                    # Device was reset, msg contains build info
-                    while not re.match(r"OpenTherm Gateway \d+(\.\d+)*", msg):
-                        msg = await self._cmdq.get()
-                    return True
-                match = re.match(expect, msg)
-                if match:
-                    if match.group(1) in v.OTGW_ERRS:
-                        # Some errors are considered a response.
-                        if retry == 0:
-                            raise v.OTGW_ERRS[match.group(1)]
-                        await send_again(msg)
-                        return
-                    ret = match.group(1)
-                    if cmd == v.OTGW_CMD_SUMMARY and ret == "1":
-                        # Expects a second line
-                        part2 = await self._cmdq.get()
-                        ret = [ret, part2]
-                    return ret
-                if re.match(r"Error 0[1-4]", msg):
-                    _LOGGER.warning(
-                        "Received %s. If this happens during a "
-                        "reset of the gateway it can be safely "
-                        "ignored.",
-                        msg,
-                    )
-                else:
-                    _LOGGER.warning("Unknown message in command queue: %s", msg)
-                await send_again(msg)
-
-            while True:
-                msg = await self._cmdq.get()
-                ret = await process(msg)
-                if ret is not None:
-                    return ret
+            _LOGGER.debug(
+                "Submitting line %d to CommandProcessor",
+                self._received_lines,
+            )
+            self.command_processor.submit_response(line)
 
     @property
     def active(self):
         """Indicate that we have seen activity on the serial line."""
-        return self._active
+        return self._received_lines > 0
 
     async def init_and_wait_for_activity(self):
         """Wait for activity on the serial connection."""
-        await self.issue_cmd(v.OTGW_CMD_SUMMARY, 0, retry=0)
+        await self.command_processor.issue_cmd(v.OTGW_CMD_SUMMARY, 0, retry=0)
         while not self.active:
             await asyncio.sleep(0)
