+++ requirements_test.txt
@@ -1,4 +1,5 @@
 bandit
+black
 flake8
 pre-commit
 pylint
