+++ tox.ini
@@ -4,7 +4,7 @@
 # and then run "tox" from this directory.
 
 [tox]
-envlist = clean, pylint, py38, py39
+envlist = clean, pylint, py38, py39, py310
 skip_missing_interpreters = True
 
 [testenv]
